+++ ChangeLog.md
@@ -19,6 +19,7 @@ Full details of the variables available for each noted event, and VoiceAttack in
       * `Location` event updated to expose new Powerplay properties.
       * `Powerplay` event updated to remove the obsolete `votes` property (Powerplay 2.0 does not use a voting system).
       * `Signal detected` event updated to include new Powerplay properties.
+      * `Power commodity fast tracked` obsolete event removed.
       * `Power defected` obsolete event removed.
       * `Power expansion vote cast` obsolete event removed.
       * `Power preparation vote cast` obsolete event removed.
@@ -27,6 +28,7 @@ Full details of the variables available for each noted event, and VoiceAttack in
       * `Jumped` script updated to include new Powerplay details.
       * `Location` script updated to include new Powerplay details.
       * `Signal detected` script updated to incorporate new Powerplay details. 
+      * `Power commodity fast tracked` obsolete script removed.
       * `Power defected` obsolete script removed.
       * `Power expansion vote cast` obsolete script removed.
       * `Power preparation vote cast` obsolete script removed.